pyproject.toml
[tool.poetry]
name = "evil_flowers_catalog"
version = "0.7.0"
description = "Simple books catalog compatible with OPDS written in Python"
authors = ["Jakub Dubec <jakub.dubec@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
django_api_forms = "1.0.0rc9"
python-dotenv = "^1.0"
argon2-cffi = "^23.1.0"
sentry-sdk = "^1.34"
psycopg = {version = "^3.1", extras = ["c"]}
django-imap-backend = "^0"
django-filter = "^23.2"
redis = "^5"
Pillow = "^10"
python-crontab = "^3"
pycountry = "^22.3"
django-object-checker = "^1.0.0"
Authlib = "^1.3"
tzdata = "^2023.4"
django-cors-headers = "^4.3"
python-ldap = "^3.4.0"
pymupdf = { version = "^1.23.2", optional = true }
minio = { version = "^7.1.12", optional = true }
qrcode = "^7.4"
bibtexparser = "^1.4.0"
isbnlib = "^3.10.14"
django-partial-date = "^1.3.2"
pydantic = "^2.5.2"
pydantic-xml = { version = "^2.7.0", extras = ["lxml"] }
email-validator = "^2.1.0.post1"

[tool.poetry.extras]
s3 = ["minio"]
pdf = ["pymupdf"]

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
datamodel-code-generator = "^0.25.2"

[tool.black]
max-line-length = 119
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.github
  | venv
  | data
  | logs
  | media
  | private
  | static
  | conf
  | \.stoplight
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.masonry.api"
